requirements.txt
python-dotenv
langchain
langchain-openai
langgraph
ipykernel
langsmith
langchain_anthropic
tavily-python
langchain-community
langchain-fireworks
bs4
chromadb
langchainhub
langchain_nomic
nomic[local]
langchain_experimental
langchain_openai
pandas
matplotlib
langchain_core
langchain-groq
mathtools
numexpr
langgraph-checkpoint-sqlite
streamlit